+++ CHANGES.rst
@@ -6,9 +6,11 @@ Version 0.2.0
 
 *Unreleased*
 
+- Add Field aliases (#14)
 - Support arbitrary serializers and deserializers (`#6`_)
 - General internal enhancements (`#5`_)
 
+.. _#14: https://github.com/rossmacarthur/serde/pull/14
 .. _#6: https://github.com/rossmacarthur/serde/pull/6
 .. _#5: https://github.com/rossmacarthur/serde/pull/5